api/pyproject.toml
[project]
name = "tanachat-api"
version = "0.1.0"
description = "TanaChat API - FastAPI backend with auth and Tana integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "boto3>=1.35.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "pyjwt>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "moto[s3]>=5.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
